Move BusinessLogic to DBAPI class
