+++ gramps/plugins/db/dbapi/sqlite.py
@@ -41,7 +41,6 @@ import sqlite3
 # -------------------------------------------------------------------------
 from gramps.gen.const import GRAMPS_LOCALE as glocale
 from gramps.gen.db.dbconst import ARRAYSIZE
-from gramps.gen.db.bizlogic import BusinessLogic
 from gramps.plugins.db.dbapi.dbapi import DBAPI
 
 _ = glocale.translation.gettext
@@ -54,7 +53,7 @@ sqlite3.paramstyle = "qmark"
 # SQLite class
 #
 # -------------------------------------------------------------------------
-class SQLite(DBAPI, BusinessLogic):
+class SQLite(DBAPI):
     """
     SQLite interface.
     """
@@ -79,51 +78,6 @@ class SQLite(DBAPI, BusinessLogic):
         else:
             path_to_db = os.path.join(directory, "sqlite.db")
         self.dbapi = Connection(path_to_db)
-
-    # -------------------------------------------------------
-    # Fast sqlite-specific implementations, slow versions
-    # in BusniessLogic. These are "underloads"
-    # -------------------------------------------------------
-    def get_father_mother_handles_from_family(self, handle=None, family=None):
-        """ Get the father and mother handles given a family """
-        if family:
-            handle = family.handle
-        self.dbapi.execute("SELECT JSON_EXTRACT(unblob, '$[2]', '$[3]') FROM family WHERE handle = ? limit 1;", [handle])
-        row = self.dbapi.fetchone()
-        if row:
-            parent_list = json.loads(row[0])
-            if parent_list:
-                return parent_list[0], parent_list[1]
-        return (None, None)
-
-    def get_main_parents_family_handle_from_person(self, handle=None, person=None):
-        """ Get the main parent's family handle given a person """
-        if person:
-            handle = person.handle
-        self.dbapi.execute("SELECT JSON_EXTRACT(unblob, '$[9]') FROM person WHERE handle = ? limit 1;", [handle])
-        row = self.dbapi.fetchone()
-        parent_family_list = json.loads(row[0])
-        if parent_family_list:
-            return parent_family_list[0]
-
-    def get_person_handle_from_gramps_id(self, gid):
-        """
-        Return the handle of the person having the given Gramps ID.
-        """
-        self.dbapi.execute("SELECT handle FROM person WHERE gramps_id = ? limit 1;", [gid])
-        row = self.dbapi.fetchone()
-        if row:
-            return row[0]
-
-    def get_father_mother_handles_from_primary_family_from_person(self, handle=None, person=None):
-        """ Get the father and mother handle's from a person primary family """
-        if person:
-            handle = person.handle
-        fam_handle = self.get_main_parents_family_handle_from_person(handle=handle)
-        if fam_handle:
-            f_handle, m_handle = self.get_father_mother_handles_from_family(handle=fam_handle)
-            return (f_handle, m_handle)
-        return (None, None)
 
 # -------------------------------------------------------------------------
 #
